重构代码生成列查询与实体属性定义
将分页参数替换为GenColumnQueryDto，支持TableId条件查询；新增GenColumnQueryDto类；优化实体属性定义，去除/添加virtual修饰符；PageResult.Total类型由long改为int；导出代码文件名调整为code.zip；生成实体名规范化为Pascal命名；统一代码风格与注释。

+++ src/Framework/Framework.WebApi/Model/PageResult.cs
@@ -10,6 +10,6 @@
         /// <summary>
         /// Gets or sets the total number of items represented by the current instance.
         /// </summary>
-        public long Total { get; set; }
+        public int Total { get; set; }
     }
 }
